+++ Game/TileMap.cpp
@@ -25,8 +25,8 @@ namespace gk
             {
                 GameObject* gameObject = ObjectFactory::Instance().Create<GameObject>(m_tileNames[index]);
 
-                float x = i % m_numX;
-                float y = i / m_numX;
+                float x = static_cast<float>(i % m_numX);
+                float y = static_cast<float>(i / m_numX);
 
                 gameObject->m_transform.position = m_start + Vector2{ x, y } * m_size;
                 scene->AddGameObject(gameObject);
